+++ app/Http/Controllers/Company/CompanyPersonController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Company;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CompanyPersonStoreRequest;
 use App\Models\Company;
+use App\Models\Permission;
 use App\Models\Person;
 use Illuminate\Http\Request;
 
@@ -12,6 +13,8 @@ class CompanyPersonController extends Controller
 {
     public function index(Company $company)
     {
+        $this->authorize('isAccess',Permission::query()->where('title','list_persons_companies')->first());
+
         return view('companyPerson.index',[
             'company'=>$company,
         ]);
@@ -19,6 +22,8 @@ class CompanyPersonController extends Controller
 
     public function create(Company $company)
     {
+        $this->authorize('isAccess',Permission::query()->where('title','create_person_company')->first());
+
         return view('companyPerson.create',[
             'company'=>$company,
             'persons'=>Person::all(),
@@ -29,6 +34,8 @@ class CompanyPersonController extends Controller
     public function store(CompanyPersonStoreRequest $request, Company $company)
     {
 //        dd($request->all(),$company);
+        $this->authorize('isAccess',Permission::query()->where('title','create_person_company')->first());
+
         $attributes=[
             'person_id' =>$request->get('person_id'),
             'startDate' =>$request->get('startDate'),
@@ -45,6 +52,8 @@ class CompanyPersonController extends Controller
     }
 
     public function destroyForm(Company $company){
+        $this->authorize('isAccess',Permission::query()->where('title','delete_person_company')->first());
+
         return view('companyPerson.destroy_form',[
             'company'=>$company,
         ]);
@@ -52,6 +61,8 @@ class CompanyPersonController extends Controller
 
     public function destroy(Company $company,Person $person)
     {
+        $this->authorize('isAccess',Permission::query()->where('title','delete_person_company')->first());
+
         $company->persons()->detach($person);
         return redirect(route('companies.persons.destroyForm',$company));
