+++ app/Http/Controllers/EducationalController.php
@@ -3,20 +3,17 @@
 namespace App\Http\Controllers;
 
 use App\Models\Educational;
+use App\Models\Permission;
 use App\Models\Person;
 use Illuminate\Http\Request;
 
 class EducationalController extends Controller
 {
 
-    public function index()
-    {
-
-    }
-
-
     public function create(Person $person)
     {
+        $this->authorize('isAccess',Permission::query()->where('title','create_educational')->first());
+
         return view('educational.create', [
             'person' => $person
         ]);
@@ -25,6 +22,8 @@ class EducationalController extends Controller
 
     public function store(Request $request, Person $person)
     {
+        $this->authorize('isAccess',Permission::query()->where('title','create_educational')->first());
+
         $educational = Educational::query()->create([
             "person_id"=>$person->id,
             "grade" => $request->get('grade'),
@@ -42,20 +41,11 @@ class EducationalController extends Controller
     }
 
 
-    public function show(Educational $educational)
-    {
-
-    }
-
-
-    public function edit(Educational $educational)
-    {
-
-    }
-
 
     public function update(Request $request, Educational $educational,Person $person)
     {
+        $this->authorize('isAccess',Permission::query()->where('title','edit_educational')->first());
+
         $educational->update([
             "person_id"=>$person->id,
             "grade" => $request->get('grade'),
@@ -76,6 +66,8 @@ class EducationalController extends Controller
 
     public function destroy(Educational $educational,Person $person)
     {
+        $this->authorize('isAccess',Permission::query()->where('title','delete_educational')->first());
+
         $educational->delete();
         return view('educational.create', [
             'person' => $person
